Split global file into several files
Having all of the code in one global file makes it difficult to parse.

It is better to split the code into separate files by purpose.

One issue is that the app may not run correctly without downloading code from a Github repo and placing it in a specific location.

In the worst case, we can upload the files to TTS, then save the generated code to the workshop mod.