Fix deck viewer rows sometimes getting misaligned
Depending on the resolution setting in Havoc TTS, that can make the card rows in Havoc TTS get misaligned, which makes reading the screen more difficult.

This fixes that issue by replacing the grid layout with a table layout so that no matter the TTS resolution setting, the cards will be aligned correctly.

Also centers the joker cards to be more visually appealing.